Fix GCC warning about non-const format string
This could be a vulnerability, as the string returned by getResult seems to be
entirely chosen by the network, and the strncmp only checks the string starts
with CLIENT_ERROR, opening way to format string vulns. Being new to this
codebase, I would however not assert it is a vulnerability without trying to
exploit it, and do not have any time to try and do it right now, so this is only
a (strong, in my opinion) possibility.

This patch keeps the hopefully intended behaviour, eliminating the possibility
for pResult to contain %-escapes.